Add types to partner policies
